requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
Pillow==7.1.2
protobuf==3.12.2
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
